requirements.txt
alembic==1.14.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
Faker==35.2.0
fastapi==0.115.5
fastapi-cli==0.0.7
fastjsonschema==2.20.0
filetype==1.2.0
Flask==3.0.3
fonttools==4.55.0
fqdn==1.5.1
fuzzywuzzy==0.18.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.5.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.7
lazy_loader==0.4
Levenshtein==0.26.1
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mysql-connector-python==9.2.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.18
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.3
openpyxl==3.1.5
orjson==3.10.13
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==11.0.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.2
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
pypiwin32==223
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==2.0.7
python-Levenshtein==0.26.1
python-multipart==0.0.17
python-socketio==5.11.4
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.10.1
readme_renderer==44.0
redis==5.2.0
referencing==0.35.1
reflex==0.6.5
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.18
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.21.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.3.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.41.3
starlette-admin==0.14.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
tkcalendar==1.6.1
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
ttkbootstrap==1.10.1
twine==5.1.1
typer==0.13.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.0
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
xlrd==2.0.1
zipp==3.21.0
